requirements.txt
requests~=2.27.1
PyYAML~=5.4.1